Introduce Suite Statements
## Motivation

In the AST, `if` nodes had a test (or condition), a body and an `orelse` representing both `elif` and `else`. The `orelse` was a `Vec<Stmt>`: One or more statements for the contents of an `else`, a single `StmtIf` for an `elif` or zero entries for neither. The problem is that this makes these two statements represented the same:

```python
if a:
    print("A")
elif b:
    print("B")
```
and
```python
if a:
    print("A")
else:
    if b:
        print("B")
```

The only way to tell them apart was to go back to the source and check the indentation of the second if statement.

## Summary

This introduces `StmtSuite`, a ranged `Vec<Stmt>` and a new statement variant, `Stmt::Suite(StmtSuite)`. Each `orlse` is now a statement, either `Stmt::If` for an elif or `Stmt::Suite` for an `else` body. Each body is now represented by the (internally identical) `StmtSuite` instead of `Vec<Stmt>`.

A lucky side effect is that the size of `Stmt` decreases (144 -> 128 bytes on x86_64). The goal is that the range of the `StmtSuite` captures the entire range from INDENT to DEDENT, this still needs to be done.

## Test Plan

Integration with the ruff formatter.